--- src/controller/ITools.java
@@ -1,13 +1,0 @@
-package controller;
-
-public interface ITools {
-
-	public void draw();
-	
-	public void fill();
-	
-	public void insert();
-	
-	public void tag();
-	
-}
