+++ src/output/awt/AWTPolyline.java
@@ -1,11 +1,12 @@
-package model.presentation.awt;
+package output.awt;
 
 import model.Shape;
-import model.presentation.ShapeState;
 import model.specific_path.Polyline;
 
 import java.awt.*;
 import java.util.ArrayList;
+
+import output.ShapeState;
 
 /**
  * Created by Xiaxing SHI on 11/11/15.
