+++ src/output/awt/AWTRectangle.java
@@ -1,10 +1,11 @@
-package model.presentation.awt;
+package output.awt;
 
 import model.Shape;
-import model.presentation.ShapeState;
 import model.specific_path.Rectangle;
 
 import java.awt.*;
+
+import output.ShapeState;
 
 /**
  * Created by Xiaxing SHI on 11/11/15.
