+++ src/output/xml/LineXML.java
@@ -1,54 +1,54 @@
-package model.presentation.xml;
+package output.xml;
 
+import output.ShapeState;
 import model.Shape;
-import model.specific_path.Text;
-import model.presentation.ShapeState;
+import model.specific_path.Line;
 
-public class TextXML implements ShapeState {
-
+public class LineXML implements ShapeState {
+	
 	/************************************
 	 *	Attributes
 	 ************************************/	
-	private Text text;
+	private Line line;
 	private StateDelegate delegate;
 	
 	/************************************
 	 *	Constructors
 	 ************************************/	
-	public TextXML(Text text, StateDelegate delegate) {
+	public LineXML(Line line, StateDelegate delegate) {
 		this.delegate = delegate;
-		this.text = text;
+		this.line = line;
 	}
 	
 	/************************************
 	 *	Getters and Setters
 	 ************************************/	
-	public Text getText() {
-		return text;
+	public Line getLine() {
+		return line;
 	}
 
-	public void setText(Text text) {
-		this.text = text;
+	public void setLine(Line line) {
+		this.line = line;
 	}
+
 	
 	/************************************
 	 *	Methods
 	 ************************************/
 	@Override
 	public String getDrawing() {
-		return "<text x=\""+this.getText().getX()+"\" y=\""+this.getText().getY()+"\" "+this.delegate.getPencilXML(this.text.getPencil())+">"+this.getText().getText()+"</text>";
+		return "<line x1=\""+this.getLine().getX1()+"\" y1=\""+this.getLine().getY1()+"\" x2=\""+this.getLine().getX2()+"\" y2=\""+this.getLine().getY2()+"\" "+this.delegate.getPencilXML(this.line.getPencil())+" />";
 	}
 	
 	@Override
 	public void setShape(Shape shape) throws Exception {
-		if(shape instanceof Text){
-			this.text = (Text) shape;
+		if(shape instanceof Line){
+			this.line = (Line) shape;
 		}
 		else{
 			throw new Exception("The shape is not of the right type");
 		}
 		
 	}
-
 
 }
